LPS-203925 Add null checking to onBeforeRemove method to avoid throwing NPE

+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/model/listener/LayoutSetModelListener.java
@@ -46,6 +46,10 @@ public class LayoutSetModelListener extends BaseModelListener<LayoutSet> {
 	public void onBeforeRemove(LayoutSet layoutSet)
 		throws ModelListenerException {
 
+		if (layoutSet == null) {
+			return;
+		}
+
 		_clientExtensionEntryRelLocalService.deleteClientExtensionEntryRels(
 			_portal.getClassNameId(LayoutSet.class),
 			layoutSet.getLayoutSetId());
